clip ctrl

+++ animal3D-SDK/animal3D-SDK/source/animal3D-DemoPlugin/A3_DEMO/a3_DemoMode0_Starter/a3_DemoMode0_Starter-load.c
@@ -173,7 +173,7 @@ void a3starter_load(a3_DemoState const* demoState, a3_DemoMode0_Starter* demoMod
 	a3keyframePoolCreate(demoMode->keyframePool, 64);
 
 	for (a3ui32 i = 0; i < 64; i++) {
-		a3keyframeInit(demoMode->keyframePool->keyframes + i, 0.5f, i);
+		a3keyframeInit(demoMode->keyframePool->keyframes + i, 0.5f, (float)i);
 	}
 
 	a3clipInit(demoMode->clipPool->clips + 0, " 0 -> 19", demoMode->keyframePool,  0, 19);
